Some assertions hit if doing it the old way, maybe this is how you're supposed to cast to an int

+++ mp/src/game/shared/neo/neo_ghost_cap_point.cpp
@@ -224,7 +224,7 @@ void CNEOGhostCapturePoint::Think_CheckMyRadius(void)
 		}
 
 		const Vector dir = player->GetAbsOrigin() - GetAbsOrigin();
-		const int distance = dir.Length();
+		const int distance = static_cast<int>(dir.Length());
 
 		Assert(distance >= 0);
